[Refactor] Move websocket handler test to jvm package
