+++ async-http-client-backend/monix/src/test/scala/sttp/client/asynchttpclient/monix/MonixWebsocketHandlerTest.scala
@@ -3,13 +3,14 @@ package sttp.client.asynchttpclient.monix
 import monix.eval.Task
 import monix.execution.Scheduler.Implicits.global
 import sttp.client._
-import sttp.client.asynchttpclient.{WebSocketHandler, WebsocketHandlerTest}
+import sttp.client.asynchttpclient.WebSocketHandler
 import sttp.client.impl.monix.{TaskMonadAsyncError, convertMonixTaskToFuture}
 import sttp.client.monad.MonadError
 import sttp.client.testing.ConvertToFuture
+import sttp.client.testing.websocket.WebsocketHandlerTest
 import sttp.client.ws.WebSocket
 
-class MonixWebsocketHandlerTest extends WebsocketHandlerTest[Task] {
+class MonixWebsocketHandlerTest extends WebsocketHandlerTest[Task, WebSocketHandler] {
   implicit val backend: SttpBackend[Task, Nothing, WebSocketHandler] = AsyncHttpClientMonixBackend().runSyncUnsafe()
   implicit val convertToFuture: ConvertToFuture[Task] = convertMonixTaskToFuture
   override implicit val monad: MonadError[Task] = TaskMonadAsyncError
