+++ async-http-client-backend/zio/src/test/scala/sttp/client/asynchttpclient/zio/ZioWebsocketHandlerTest.scala
@@ -1,14 +1,15 @@
 package sttp.client.asynchttpclient.zio
 
 import sttp.client._
-import sttp.client.asynchttpclient.{WebSocketHandler, WebsocketHandlerTest}
+import sttp.client.asynchttpclient.WebSocketHandler
 import sttp.client.impl.zio.{TaskMonadAsyncError, convertZioIoToFuture, runtime}
 import sttp.client.monad.MonadError
 import sttp.client.testing.ConvertToFuture
+import sttp.client.testing.websocket.WebsocketHandlerTest
 import sttp.client.ws.WebSocket
 import zio.Task
 
-class ZioWebsocketHandlerTest extends WebsocketHandlerTest[Task] {
+class ZioWebsocketHandlerTest extends WebsocketHandlerTest[Task, WebSocketHandler] {
   override implicit val backend: SttpBackend[Task, Nothing, WebSocketHandler] =
     runtime.unsafeRun(AsyncHttpClientZioBackend())
   override implicit val convertToFuture: ConvertToFuture[Task] = convertZioIoToFuture
